修改数据类型，避免短信分片的pkNumber ,pkTotle,pkSeq是负数

+++ src/main/java/com/zx/sms/codec/cmpp/wap/LongMessageFrame.java
@@ -15,9 +15,9 @@ public class LongMessageFrame implements Serializable{
 	private static final long serialVersionUID = -8554060199834235624L;
 	
 	private long timestamp;
-	private short pkseq = 0;
-	private byte pktotal = 1;
-	private byte pknumber = 1;
+	private int pkseq = 0;
+	private short pktotal = 1;
+	private short pknumber = 1;
 	private short tppid = 0;// 0是普通GSM 类型，点到点方式 ,127 :写sim卡
 	private short tpudhi = 0; // 0:msgcontent不带协议头。1:带有协议头
 	private AbstractSmsDcs msgfmt = GlobalConstance.defaultmsgfmt;
@@ -26,18 +26,18 @@ public class LongMessageFrame implements Serializable{
 
 	private int sequence;
 	
-	public short getPkseq() {
+	public int getPkseq() {
 		return pkseq;
 	}
 
-	public void setPkseq(short pkseq) {
+	public void setPkseq(int pkseq) {
 		this.pkseq = pkseq;
 	}
 
 	/**
 	 * @return the pktotal
 	 */
-	public byte getPktotal() {
+	public short getPktotal() {
 		return pktotal;
 	}
 
@@ -45,14 +45,14 @@ public class LongMessageFrame implements Serializable{
 	 * @param pktotal
 	 *            the pktotal to set
 	 */
-	public void setPktotal(byte pktotal) {
+	public void setPktotal(short pktotal) {
 		this.pktotal = pktotal;
 	}
 
 	/**
 	 * @return the pknumber
 	 */
-	public byte getPknumber() {
+	public short getPknumber() {
 		return pknumber;
 	}
 
@@ -60,7 +60,7 @@ public class LongMessageFrame implements Serializable{
 	 * @param pknumber
 	 *            the pknumber to set
 	 */
-	public void setPknumber(byte pknumber) {
+	public void setPknumber(short pknumber) {
 		this.pknumber = pknumber;
 	}
